+++ .env
@@ -1,0 +1,8 @@
+export const CLIENT_ID = "8dfa7db4e6e6e29a50acefe5f2016a73";
+export const REACT_APP_REST_API_KEY = "8dfa7db4e6e6e29a50acefe5f2016a73";
+export const REACT_APP_KAKAO_REST_KEY = "8dfa7db4e6e6e29a50acefe5f2016a73";
+
+export const REDIRECT_URI =  "https://localhost:3000/auth/kakao";
+export const REACT_APP_REDIRECT_URL = "https://www.match-api-server/auth/kakao";
+
+export const KAKAO_AUTH_URL = `https://kauth.kakao.com/oauth/authorize?client_id=${CLIENT_ID}&redirect_uri=${REDIRECT_URI}&response_type=code`;
