--- src/components/KakaoLogin/KakaoLogin.ts
@@ -1,9 +1,0 @@
-
-
-
-const KakaoLogin = () => {
-
-}
-
-
-export default KakaoLogin
