+++ src/react-app-env.d.ts
@@ -1,3 +1,5 @@
 /// <reference types="react-scripts" />
-const REACT_APP_KAKAO_REST_KEY = "8dfa7db4e6e6e29a50acefe5f2016a73";
 
+interface Window {
+    Kakao: any
+}
